Refactor mappers and services for Administrador and Terapeuta entities
- Updated method names in AdministradorMapper for clarity.
- Added new methods in TerapeutaMapper for DTO conversions.
- Changed password field name from 'contrasenia' to 'contrasena' in multiple DTOs and entities.
- Implemented password reset functionality in TerapeutaService and added corresponding endpoint in TerapeutaController.
- Created new AdministradorController for handling administrador creation.
- Adjusted service methods to use new DTOs and handle password encoding.
- Enhanced security configuration to permit access to administrador endpoints.
- Updated repository methods to reflect changes in entity relationships and types.

+++ back-end/back/src/main/java/com/remind/back/repositories/TerapeutaRepository.java
@@ -3,14 +3,13 @@ package com.remind.back.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.remind.back.entities.Terapeuta;
 import java.util.Optional;
-import java.util.List;
 
 
 public interface TerapeutaRepository extends JpaRepository<Terapeuta, Integer> {
     
     Optional<Terapeuta> findByEmail(String email);
 
-    Optional<Terapeuta> findById(Integer id);
+    Optional<Terapeuta> findById(Long id);
 
     Optional<Terapeuta> findByUsuario(String usuario);
 }
